pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "google-adk-redis"
version = "0.1.1"
description = "Redis-backed session service for Google ADK"
readme = "README.md"
requires-python = ">=3.9"
license = { file = "LICENSE" }
authors = [
  { name = "BloodBoy21" },
  { name = "hucruz" },
]
maintainers = [
  { name = "hucruz" },
]
dependencies = [
  "google-adk>=1.10.0",
  "redis>=6.0.0",
  "typing-extensions>=4.5,<5",
]

[project.optional-dependencies]
dev = [
  "pytest>=7.0",
  "pytest-asyncio>=0.21",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.setuptools.packages.find]
where = ["src"]
